+++ relay-server/src/actors/processor.rs
@@ -2337,6 +2337,15 @@ impl EnvelopeProcessorService {
         if all_modules_enabled {
             // Extract tags to add to this span as well
             let shared_tags = span::tag_extraction::extract_shared_tags(event);
+            transaction_span.sentry_tags = Annotated::new(
+                shared_tags
+                    .clone()
+                    .into_iter()
+                    .map(|(k, v)| (k.sentry_tag_key().to_owned(), Annotated::new(v)))
+                    .collect(),
+            );
+            // Double write to `span.data` for now. This can be removed once all users of these fields
+            // have switched to `sentry_tags`.
             let data = transaction_span
                 .data
                 .value_mut()
@@ -2345,7 +2354,7 @@ impl EnvelopeProcessorService {
                 shared_tags
                     .clone()
                     .into_iter()
-                    .map(|(k, v)| (k.to_string(), Annotated::new(v))),
+                    .map(|(k, v)| (k.data_key().to_owned(), Annotated::new(v.into()))),
             );
             add_span(transaction_span.into());
         }